+++ R/tables.R
@@ -10,20 +10,20 @@
 #' @param sort if specified, allow to sort the table by increasing ("inc") or decreasing ("dec") frequencies
 #' @param valid if TRUE, display valid percentages
 #' @param levels the desired levels for the factor in case of labelled vector (\pkg{labelled} package
-#'    must be installed): "labels" for value labels, "values" for values or 
+#'    must be installed): "labels" for value labels, "values" for values or
 #'    "prefixed" for labels prefixed with values
 #' @param na.last if TRUE, NA values are always be last table row
 #' @return
 #' The result is an object of class data.frame.
 #' @seealso
 #' \code{\link{table}}, \code{\link[questionr]{prop}}, \code{\link[questionr]{cprop}}, \code{\link[questionr]{rprop}}
-#' @examples 
+#' @examples
 #' # factor
 #' data(hdv2003)
 #' freq(hdv2003$qualif)
 #' freq(hdv2003$qualif, cum = TRUE, total = TRUE)
 #' freq(hdv2003$qualif, cum = TRUE, total = TRUE, sort ="dec")
-#' 
+#'
 #' # labelled data
 #' data(fecondite)
 #' freq(femmes$region)
@@ -31,22 +31,22 @@
 #' freq(femmes$region, levels = "v")
 #' @export
 
-freq <- function(x, digits = 1, cum = FALSE, total = FALSE, exclude = NULL, sort = "", 
+freq <- function(x, digits = 1, cum = FALSE, total = FALSE, exclude = NULL, sort = "",
          valid = !(NA %in% exclude), levels = c("prefixed", "labels", "values"),
          na.last = TRUE) {
-  
+
   levels <- match.arg(levels)
-  
+
   if (is.table(x)) {
     tab <- x
   } else {
     tab <- table(labelled::to_factor(x, levels), exclude = exclude)
   }
-  
+
   effectifs <- as.vector(tab)
   pourc <- as.vector(effectifs / sum(effectifs) * 100)
   result <- data.frame(n = effectifs, pourc = pourc)
-  
+
   if (valid) {
     user_na <- unique(as.character(labelled::to_factor(x, levels)[is.na(x)]))
     NA.position <- which(is.na(names(tab)) | names(tab) %in% user_na)
@@ -55,24 +55,24 @@ freq <- function(x, digits = 1, cum = FALSE, total = FALSE, exclude = NULL, sort
     valid.pourc[NA.position] <- 0 # temporary 0 for cumsum
     result <- cbind(result, valid.pourc)
   }
-  
+
   ## Avoid duplicate row names if both NA and "NA" in tab
   if ("NA" %in% names(tab)) {
     names(tab)[names(tab) == "NA"] <- "\"NA\""
   }
   rownames(result) <- ifelse(is.na(names(tab)), "NA", names(tab))
-  
+
   if (sort == "inc") result <- result[order(result$n),]
   if (sort == "dec") result <- result[order(result$n, decreasing = TRUE),]
-  
+
   if (na.last && "NA" %in% rownames(result)) {
     result <- rbind(result[-which(rownames(result) == "NA"), ], result["NA", ])
   }
-  
+
   if (total) result <- rbind(result, Total = apply(result,2,sum))
-  if (total & valid) 
+  if (total & valid)
     result[length(result$pourc),"valid.pourc"] <- 100
-  
+
   if (cum) {
     pourc.cum <- cumsum(result$pourc)
     if (total) pourc.cum[length(pourc.cum)] <- 100
@@ -83,21 +83,21 @@ freq <- function(x, digits = 1, cum = FALSE, total = FALSE, exclude = NULL, sort
       result <- cbind(result, valid.pourc.cum)
     }
   }
-  
+
   if (valid) {
     NA.position <- which(rownames(result) == "NA" | rownames(result) %in% user_na)
     result[NA.position, "valid.pourc"] <- NA
     if (cum)
       result[NA.position, "valid.pourc.cum"] <- NA
   }
-  
+
   names(result)[names(result) == "pourc"] <- "%"
   names(result)[names(result) == "valid.pourc"] <- "val%"
   names(result)[names(result) == "pourc.cum"] <- "%cum"
   names(result)[names(result) == "valid.pourc.cum"] <- "val%cum"
-  
+
   class(result) <- c("freqtab", class(result))
-  
+
   round(result, digits = digits)
 }
 
@@ -133,10 +133,10 @@ freq.na <- function(data, ...) {
     d = as.data.frame(data)
   }
   if (is.null(dim(d))) {
-    c = length(d)  
+    c = length(d)
   }
   else {
-    c = nrow(d)    
+    c = nrow(d)
   }
   d = is.na(as.matrix(d))
   d = as.matrix(colSums(d))
@@ -147,19 +147,19 @@ freq.na <- function(data, ...) {
     names(d) = n
   else
     colnames(d) = n
-  
+
   return(d)
 }
 
-#' Column percentages of a two-way frequency table.
+#' Column percentages of a cross-tabulation table (2 dimensions or more).
 #'
-#' Return the column percentages of a two-way frequency table with formatting and printing options.
+#' Return the column percentages of a cross-tabulation table (2 dimensions or more) with formatting and printing options.
 #'
 #' @param tab frequency table
 #' @param digits number of digits to display
 #' @param total if \code{TRUE}, add a row with the sum of percentages and a column with global percentages
 #' @param percent if \code{TRUE}, add a percent sign after the values when printing
-#' @param drop if \code{TRUE}, lines or columns with a sum of zero, which would generate \code{NaN} percentages, are dropped.
+#' @param drop if \code{TRUE}, lines or columns with a sum of zero, which would generate \code{NaN} percentages, are dropped.  Unused for tables of 3 dimensions or more (always `FALSE`).
 #' @param n if \code{TRUE}, display number of observations per column.
 #' @param ... parameters passed to other methods.
 #' @return
@@ -174,6 +174,8 @@ freq.na <- function(data, ...) {
 #' cprop(tab)
 #' ## Column percentages with custom display
 #' cprop(tab, digits=2, percent=TRUE, total=FALSE)
+#' ## Could be applied to a table of 3 dimensions or more
+#' cprop(Titanic)
 #' @export
 
 `cprop` <- function(tab, ...) {
@@ -185,7 +187,34 @@ freq.na <- function(data, ...) {
 ##' @aliases cprop.table
 ##' @export
 
-cprop.table <- function (tab, digits = 1, total = TRUE, percent = FALSE, drop = TRUE, n=FALSE, ...) {
+cprop.table <- function (tab, digits = 1, total = TRUE, percent = FALSE, drop = TRUE, n = FALSE, ...) {
+  # 3 dimensions or more
+  if (length(dim(tab)) > 2) {
+    r <- apply(
+      tab,
+      3:length(dim(tab)),
+      cprop,
+      simplify = FALSE,
+      digits = digits,
+      total = total,
+      percent = percent,
+      drop = FALSE,
+      n = n,
+      ...
+    )
+    res <- array(
+      unlist(r),
+      dim = c(dim(r[[1]]), dim(tab)[-2:-1]),
+      dimnames = append(dimnames(r[[1]]), dimnames(tab)[-2:-1])
+    )
+    class(res) <- class(r[[1]])
+    attr(res, "percent") <- attr(r[[1]], "percent")
+    attr(res, "digits") <- attr(r[[1]], "digits")
+    attr(res, "total") <- attr(r[[1]], "total")
+    attr(res, "row.n") <- attr(r[[1]], "row.n")
+    return(res)
+  }
+
   # subset to non-empty rows/columns
   if(drop) tab <- tab[rowSums(tab) > 0, colSums(tab) > 0, drop=FALSE]
   dn <- names(dimnames(tab))
@@ -214,11 +243,11 @@ cprop.table <- function (tab, digits = 1, total = TRUE, percent = FALSE, drop = 
 
 ##' @rdname cprop
 ##' @aliases cprop.data.frame
-##' @export 
+##' @export
 cprop.data.frame <- cprop.table
 ##' @rdname cprop
 ##' @aliases cprop.matrix
-##' @export 
+##' @export
 cprop.matrix <- cprop.table
 
 ##' @rdname cprop
@@ -239,16 +268,16 @@ cprop.tabyl <- function(tab, digits = 1, total = TRUE, percent = FALSE, n = FALS
 }
 
 
-#' Row percentages of a two-way frequency table.
+#' Row percentages of a cross-tabulation table (2 dimensions or more).
 #'
-#' Return the row percentages of a two-way frequency table with formatting and printing options.
+#' Return the row percentages of a cross-tabulation table (2 dimensions or more) with formatting and printing options.
 #'
 #' @aliases lprop
 #' @param tab frequency table
 #' @param digits number of digits to display
 #' @param total if \code{TRUE}, add a column with the sum of percentages and a row with global percentages
 #' @param percent if \code{TRUE}, add a percent sign after the values when printing
-#' @param drop if \code{TRUE}, lines or columns with a sum of zero, which would generate \code{NaN} percentages, are dropped.
+#' @param drop if \code{TRUE}, lines or columns with a sum of zero, which would generate \code{NaN} percentages, are dropped. Unused for tables of 3 dimensions or more (always `FALSE`).
 #' @param n if \code{TRUE}, display number of observations per row.
 #' @param ... parameters passed to other methods.
 #' @return
@@ -263,6 +292,8 @@ cprop.tabyl <- function(tab, digits = 1, total = TRUE, percent = FALSE, n = FALS
 #' rprop(tab)
 #' ## Column percentages with custom display
 #' rprop(tab, digits=2, percent=TRUE, total=FALSE)
+#' ## Could be applied to a table of 3 dimensions or more
+#' rprop(Titanic)
 #' @export rprop lprop
 
 `rprop` <- function(tab, ...) {
@@ -273,8 +304,35 @@ lprop <- rprop
 
 ##' @rdname rprop
 ##' @aliases rprop.table
-##' @export 
-rprop.table <- function(tab, digits = 1, total = TRUE, percent = FALSE, drop = TRUE, n=FALSE, ...) {
+##' @export
+rprop.table <- function(tab, digits = 1, total = TRUE, percent = FALSE, drop = TRUE, n = FALSE, ...) {
+  # 3 dimensions or more
+  if (length(dim(tab)) > 2) {
+    r <- apply(
+      tab,
+      3:length(dim(tab)),
+      rprop,
+      simplify = FALSE,
+      digits = digits,
+      total = total,
+      percent = percent,
+      drop = FALSE,
+      n = n,
+      ...
+    )
+    res <- array(
+      unlist(r),
+      dim = c(dim(r[[1]]), dim(tab)[-2:-1]),
+      dimnames = append(dimnames(r[[1]]), dimnames(tab)[-2:-1])
+    )
+    class(res) <- class(r[[1]])
+    attr(res, "percent") <- attr(r[[1]], "percent")
+    attr(res, "digits") <- attr(r[[1]], "digits")
+    attr(res, "total") <- attr(r[[1]], "total")
+    attr(res, "col.n") <- attr(r[[1]], "col.n")
+    return(res)
+  }
+
   # subset to non-empty rows/columns
   if(drop) tab <- tab[rowSums(tab) > 0, colSums(tab) > 0, drop=FALSE]
   dn <- names(dimnames(tab))
@@ -302,11 +360,11 @@ rprop.table <- function(tab, digits = 1, total = TRUE, percent = FALSE, drop = T
 }
 ##' @rdname rprop
 ##' @aliases rprop.data.frame
-##' @export 
+##' @export
 rprop.data.frame <- rprop.table
 ##' @rdname rprop
 ##' @aliases rprop.matrix
-##' @export 
+##' @export
 rprop.matrix <- rprop.table
 
 ##' @rdname rprop
@@ -325,18 +383,18 @@ rprop.tabyl <- function(tab, digits = 1, total = TRUE, percent = FALSE, n = FALS
   tab <- janitor::adorn_title(tab, "combined")
   return(tab)
 }
- 
 
 
-#' Global percentages of a two-way frequency table.
+
+#' Global percentages of a cross-tabulation table (2 dimensions or more).
 #'
-#' Return the percentages of a two-way frequency table with formatting and printing options.
+#' Return the percentages of a cross-tabulation table (2 dimensions or more) with formatting and printing options.
 #'
 #' @param tab frequency table
 #' @param digits number of digits to display
 #' @param total if \code{TRUE}, add a column with the sum of percentages and a row with global percentages
 #' @param percent if \code{TRUE}, add a percent sign after the values when printing
-#' @param drop if \code{TRUE}, lines or columns with a sum of zero, which would generate \code{NaN} percentages, are dropped.
+#' @param drop if \code{TRUE}, lines or columns with a sum of zero, which would generate \code{NaN} percentages, are dropped. Unused for tables of 3 dimensions or more (always `FALSE`).
 #' @param n if \code{TRUE}, display number of observations per row and per column.
 #' @param ... parameters passed to other methods
 #' @return
@@ -351,6 +409,8 @@ rprop.tabyl <- function(tab, digits = 1, total = TRUE, percent = FALSE, n = FALS
 #' prop(tab)
 #' ## Percentages with custom display
 #' prop(tab, digits=2, percent=TRUE, total=FALSE, n=TRUE)
+#' ## Could be applied to a table of 3 dimensions or more
+#' prop(Titanic)
 #' @export
 
 prop <- function(tab, ...) {
@@ -363,9 +423,37 @@ prop <- function(tab, ...) {
 
 ##' @rdname prop
 ##' @aliases prop_table
-##' @export 
+##' @export
 
-prop_table <- function (tab, digits = 1, total = TRUE, percent = FALSE, drop = TRUE, n=FALSE, ...) {
+prop_table <- function (tab, digits = 1, total = TRUE, percent = FALSE, drop = TRUE, n = FALSE, ...) {
+  # 3 dimensions or more
+  if (length(dim(tab)) > 2) {
+    r <- apply(
+      tab,
+      3:length(dim(tab)),
+      prop_table,
+      simplify = FALSE,
+      digits = digits,
+      total = total,
+      percent = percent,
+      drop = FALSE,
+      n = n,
+      ...
+    )
+    res <- array(
+      unlist(r),
+      dim = c(dim(r[[1]]), dim(tab)[-2:-1]),
+      dimnames = append(dimnames(r[[1]]), dimnames(tab)[-2:-1])
+    )
+    class(res) <- class(r[[1]])
+    attr(res, "percent") <- attr(r[[1]], "percent")
+    attr(res, "digits") <- attr(r[[1]], "digits")
+    attr(res, "total") <- attr(r[[1]], "total")
+    attr(res, "row.n") <- attr(r[[1]], "row.n")
+    attr(res, "col.n") <- attr(r[[1]], "col.n")
+    return(res)
+  }
+
   # subset to non-empty rows/columns
   if(drop) tab <- tab[rowSums(tab) > 0, colSums(tab) > 0, drop=FALSE]
   dn <- names(dimnames(tab))
@@ -400,11 +488,11 @@ prop_table <- function (tab, digits = 1, total = TRUE, percent = FALSE, drop = T
 
 ##' @rdname prop
 ##' @aliases prop.data.frame
-##' @export 
+##' @export
 prop.data.frame <- prop_table
 ##' @rdname prop
 ##' @aliases prop.matrix
-##' @export 
+##' @export
 prop.matrix <- prop_table
 
 ##' @rdname prop
@@ -428,7 +516,7 @@ prop.tabyl <- function(tab, digits = 1, total = TRUE, percent = FALSE, n = FALSE
 
 #' Return the chi-squared residuals of a two-way frequency table.
 #'
-#' Return the raw, standardized or Pearson's residuals (the default) of a chi-squared test on a two-way frequency table. 
+#' Return the raw, standardized or Pearson's residuals (the default) of a chi-squared test on a two-way frequency table.
 #'
 #' @aliases residus
 #' @param tab frequency table
@@ -496,18 +584,27 @@ format.proptab <- function (x, digits=NULL, percent=NULL, justify="right", ...) 
   row.n <- attr(x, "row.n"); if (is.null(row.n)) row.n <- FALSE
   col.n <- attr(x, "col.n"); if (is.null(col.n)) col.n <- FALSE
   tmp <- format.default(round(x,0), ...)
-  if (row.n) rn <- tmp[nrow(x),]
-  if (col.n) cn <- tmp[,ncol(x)]
+  if (row.n) {
+    posr <- dim(x)[1] * 1:(length(x) / dim(x)[1]) # positions of the n for rows
+    rn <- tmp[posr]
+  }
+  if (col.n) {
+    posc1 <- dim(x)[1] * (dim(x)[2] - 1) + seq_len(dim(x)[1]) # first table
+    ntable <- length(x) / (dim(x)[1] * dim(x)[2]) # number of tables
+    posc <- c() # positions of the n for cols
+    for (i in seq_len(ntable)) posc <- c(posc, posc1 + (i - 1) * dim(x)[1] * dim(x)[2])
+    cn <- tmp[posc]
+  }
   if (percent) {
     fmt <- paste("%.",digits,"f%%",sep="")
-    x[,] <- sprintf(x, fmt=fmt)
+    x[] <- sprintf(x, fmt=fmt)
     result <- format.default(x,justify=justify, ...)
   }
   else
     result <- format.default(round(x,digits), ...)
-  if (row.n) result[nrow(x),] <- rn
-  if (col.n) result[,ncol(x)] <- cn
-  if (total & row.n & col.n) result[nrow(x),ncol(x)] <- ""
+  if (row.n) result[posr] <- rn
+  if (col.n) result[posc] <- cn
+  if (total && row.n && col.n) result[intersect(posc, posr)] <- ""
   return(result)
 }
 
@@ -532,19 +629,19 @@ print.proptab <- function (x, digits=NULL, percent=NULL, justify="right", ...) {
 
 
 #' Cross tabulation with labelled variables
-#' 
+#'
 #' This function is a wrapper around \code{\link[stats]{xtabs}}, adding automatically
 #' value labels for labelled vectors if \pkg{labelled} package eis installed.
-#' 
+#'
 #' @param formula a formula object (see \code{\link[stats]{xtabs}})
 #' @param data a data frame
-#' @param levels the desired levels in case of labelled vector: 
+#' @param levels the desired levels in case of labelled vector:
 #'    "labels" for value labels, "values" for values or "prefixed" for labels prefixed with values
 #' @param variable_label display variable label if available?
 #' @param ... additional arguments passed to \code{\link[stats]{xtabs}}
-#' 
+#'
 #' @seealso \code{\link[stats]{xtabs}}.
-#' @examples 
+#' @examples
 #' data(fecondite)
 #' ltabs(~radio, femmes)
 #' ltabs(~radio+tv, femmes)
@@ -555,34 +652,34 @@ print.proptab <- function (x, digits=NULL, percent=NULL, justify="right", ...) {
 #' @export
 #' @importFrom stats as.formula
 #' @importFrom stats terms
-#' @importFrom stats xtabs 
+#' @importFrom stats xtabs
 
 ltabs <- function(formula, data, levels = c("prefixed", "labels", "values"), variable_label = TRUE, ...){
     levels <- match.arg(levels)
     formula <- stats::as.formula(formula)
     if (!is.data.frame(data))
       data <- as.data.frame(data)
-    
+
     vars <- attr(stats::terms(formula), "term.labels")
-    
+
     dn <- vars
     for (i in 1:length(vars))
       if (!is.null(labelled::var_label(data[[vars[i]]])) & variable_label)
         dn[i] <- paste(vars[i], labelled::var_label(data[[vars[i]]]), sep = ": ")
-    
-    for (v in vars) 
+
+    for (v in vars)
       data[[v]] <- labelled::to_factor(data[[v]], levels = levels)
-    
+
     tab <- stats::xtabs(formula, data, ...)
     names(dimnames(tab)) <- dn
     return(tab)
   }
 
 #' Frequency table of variables
-#' 
+#'
 #' Generate frequency tables for one or more variables in a data frame
 #' or a survey design.
-#' 
+#'
 #' @aliases freqtable.default freqtable.survey.design
 #' @param .data a data frame or `survey.design` object
 #' @param ... one or more expressions accepted by \code{\link[dplyr]{select}}
@@ -593,22 +690,22 @@ ltabs <- function(formula, data, levels = c("prefixed", "labels", "values"), var
 #' If `.data` is a survey design, either `TRUE` (the default) to
 #' to use survey weights, or `FALSE` or `NULL` to return unweighted
 #' frequencies.
-#' 
+#'
 #' @return The result is an array of class `table`.
 #' @seealso \code{\link{freq}}, \code{\link{wtd.table}},
 #'          \code{\link{xtabs}}, \code{\link{table}}.
-#' @examples 
+#' @examples
 #' data(hdv2003)
 #' freqtable(hdv2003, nivetud, sport)
 #' freqtable(hdv2003, nivetud, sport, sexe)
 #' freqtable(hdv2003, nivetud, sport, weights=poids)
 #' freqtable(hdv2003, starts_with("trav"))
-#' 
+#'
 #' # Using survey design objects
 #' library(survey)
 #' hdv2003_wtd <- svydesign(ids=~1, weights=~poids, data=hdv2003)
 #' freqtable(hdv2003_wtd, nivetud, sport)
-#' 
+#'
 #' # Compute percentages based on frequencies
 #' hdv2003 |> freqtable(sport) |> freq()
 #' hdv2003 |> freqtable(sport, sexe) |> prop()