+++ NEWS.md
@@ -1,6 +1,7 @@
 # questionr (development version)
 
 * `irec()` and `iorder()` have been updated to use the native pipe and `fct_na_value_to_level()` instead of `fct_explicit_na()` (#141)
+* `prop()`, `cprop()` and `rprop()` could now be applied on tables of 3 dimensions or more (#38)
 
 # questionr 0.7.8