+++ NEWS.md
@@ -1,4 +1,4 @@
-# questionr (development version)
+# questionr 0.8.0
 
 * `irec()` and `iorder()` have been updated to use the native pipe and `fct_na_value_to_level()` instead of `fct_explicit_na()` (#141)
 * `prop()`, `cprop()` and `rprop()` could now be applied on tables of 3 dimensions or more (#38)